requirements.txt
gitpython>=1.7
